pyproject.toml
[project]
name = "scott_tracy"
version = "0.0.1"
description = "Pilot of the swing-wing rocket-plane Thunderbird 1."
authors = [ { name = "Joe J Collins", email = "joejcollins@gmail.com" },]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
requires-python = ">=3.12"
dependencies = [
    "debugpy",
    "ipykernel",
    "ipympl",
    "loguru",
    "mkdocs",
    "pre-commit",
    "python-dotenv",
    "pytest",
    "pytest-cov",
    "ruff",
    "sourcery",
    "tqdm",
    "types-tqdm",
]

[project.urls]
repository = "https://github.com/earthroverprogram/scott-tracy"

[tool.pytest.ini_options]
log_cli = false
addopts = "-q --disable-warnings --cache-clear"
testpaths = ["./tests"]

[tool.ruff]
lint.select = ["F", "E", "W", "C90", "I", "PL", "D"]
lint.ignore = ["D203", "D213"]
lint.fixable = ["ALL"]
line-length = 88

[tool.uv]
package = true

[[tool.mypy.overrides]]
module = ["scott_tracy.*"]
follow_untyped_imports = true

[tool.pyright]
exclude = ["**/__pycache__"]
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.13"
pythonPlatform = "Linux"
